# Conflicts:
#	build.gradle
#	common/src/main/java/io/github/apace100/origins/access/WaterMovingEntity.java
#	common/src/main/java/io/github/apace100/origins/command/OriginCommand.java
#	common/src/main/java/io/github/apace100/origins/power/factory/PowerFactories.java
#	common/src/main/resources/assets/origins/lang/es_ar.json
#	common/src/main/resources/assets/origins/lang/es_cl.json
#	common/src/main/resources/assets/origins/lang/es_ec.json
#	common/src/main/resources/assets/origins/lang/es_mx.json
#	common/src/main/resources/assets/origins/lang/es_uy.json
#	common/src/main/resources/assets/origins/lang/es_ve.json
#	fabric/src/main/java/io/github/apace100/origins/mixin/fabric/CapeFeatureRendererMixin.java
#	gradle.properties
#	src/main/resources/origins.mixins.json

+++ common/src/main/java/io/github/apace100/origins/power/factory/PowerFactories.java
@@ -336,7 +336,7 @@ public class PowerFactories {
                         (type, player) ->
                                 new ModifyPlayerSpawnPower(type, player,
                                         data.get("dimension"),
-                                        (int)data.getFloat("dimension_distance_multiplier"),
+                                        data.getFloat("dimension_distance_multiplier"),
                                         data.get("biome"),
                                         data.get("spawn_strategy"),
                                         data.isPresent("structure") ? data.get("structure") : null,
